+++ src/main/java/com/vti/repository/ICategoryRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.vti.entity.Category;
 
-public interface ICategoryRepository extends JpaRepository<Category, Short> {
+public interface ICategoryRepository extends JpaRepository<Category, Integer> {
 	public Category findByName(String name);
 
 	public boolean existsByName(String name);
